moved the nullcharremoval test to functional dir
